Improve dashboard UI
On the dashboard page, added loading animation and animation to all
cards when they appear. Changed the top 3 cards to have an image. Added
more information to the recent news section. Also added the view more
button.

Added animation to news page.

Added animation to news article page and changed the layout to card
over.

In the news creator, added animation to when the page first loads.
Changed the stepper to be vertical and to follow the teradata
recommendations. Added a loading animation when submitting and then a
submit message.

Resolves #16